Update API_Invoke.md
Invocations occur at least once in response to an Event and functions must be idempotent to handle this
Can be seen as ambiguous as the least once policy applies to InvocationType=event not RequestResponse. 
It’s also import to note the difference as error handling differs based on this information.